Configures SonarQube for pull request analysis
Adds configuration to the SonarQube scanner to enable
pull request analysis. This allows SonarQube to
associate issues with the pull request and display
them in the pull request UI.

Also, this commit modifies the build process to create a
dedicated build directory before running CMake,
and adjusts the order of build-wrapper execution.